align HostInformation with libogc hostent ABI
libogc's netdb.h defines struct hostent::h_addrtype and
struct hostent::h_length as C 'short' members. Accordingly,
bindgen correctly generated these fields as i16. The safe Rust
HostInformation wrapper instead declared both fields as i32,
causing compilation failures when copying values directly from
the FFI structure. Thus, I was getting this error:

error[E0308]: mismatched types
   --> src\network.rs:294:31
    |
294 |                 address_type: (*r).h_addrtype,
    |                               ^^^^^^^^^^^^^^^ expected i32, found i16

error[E0308]: mismatched types
   --> src\network.rs:295:25
    |
295 |                 length: (*r).h_length,
    |                         ^^^^^^^^^^^^^ expected i32, found i16

The update makes HostInformation to match the underlying C ABI to fix an error
and have the wrapper accurately reflects libogc's networking interface
without requiring a conversion.

libogc netdb.h:
  https://github.com/devkitPro/libogc/blob/master/external/bsd/include/netdb.h

+++ src/network.rs
@@ -210,9 +210,9 @@ pub struct HostInformation {
     /// A NULL-terminated array of alternate names.
     pub aliases: Vec<String>,
     /// The type of address being returned.
-    pub address_type: i32,
+    pub address_type: i16,
     /// The length, in bytes, of each address.
-    pub length: i32,
+    pub length: i16,
     /// A NULL-terminated list of addresses for the host.
     pub address_list: Vec<String>,
 }
